basic Room class
